requirements.txt
msrestazure>=0.4.14
azure-mgmt-resource>=1.1.0
azure-keyvault>=4.0.0
azure-identity
